# Improve VictoryTooltip rendering performance